split mod specific parts of legacy map import
